Create simple module
