+++ pyqir/src/generator.rs
@@ -18,23 +18,20 @@ use crate::{
     context::{self, Context},
     instructions::IntPredicate,
     types::Type,
-    utils::{
-        any_to_meta, call_if_some, clone_module, extract_constant, function_type,
-        try_callable_value,
-    },
+    utils::{any_to_meta, call_if_some, extract_constant, try_callable_value},
     values::{self, Value},
 };
 use inkwell::{
     attributes::Attribute as InkwellAttribute, builder::Builder as InkwellBuilder,
     context::Context as InkwellContext, memory_buffer::MemoryBuffer,
-    module::Module as InkwellModule, types::AnyTypeEnum, values::IntValue, values::PointerValue,
+    module::Module as InkwellModule, values::IntValue, values::PointerValue,
 };
 use pyo3::{
     exceptions::{PyOSError, PyValueError},
     prelude::*,
     types::{PyBytes, PySequence, PyString, PyUnicode},
 };
-use qirlib::{module, types, BuilderBasicQisExt};
+use qirlib::{module, BuilderBasicQisExt};
 use std::{
     convert::{Into, TryInto},
     mem::transmute,
@@ -93,7 +90,7 @@ impl Module {
 }
 
 impl Module {
-    fn new(py: Python, context: Py<Context>, name: &str) -> Self {
+    pub(crate) fn new(py: Python, context: Py<Context>, name: &str) -> Self {
         let module = {
             let context = context.borrow(py);
             let module = context.create_module(name);
@@ -101,106 +98,13 @@ impl Module {
         };
         Self { module, context }
     }
-}
 
-/// Provides access to all supported types.
-#[pyclass]
-pub(crate) struct TypeFactory {
-    module: Py<Module>,
-}
-
-#[pymethods]
-impl TypeFactory {
-    /// The void type.
-    ///
-    /// :type: Type
-    #[getter]
-    fn void(&self, py: Python) -> PyResult<PyObject> {
-        self.create_type(py, |m| m.get_context().void_type().into())
+    pub(crate) fn get(&self) -> &InkwellModule<'static> {
+        &self.module
     }
 
-    /// The boolean type.
-    ///
-    /// :type: Type
-    #[getter]
-    fn bool(&self, py: Python) -> PyResult<PyObject> {
-        self.create_type(py, |m| m.get_context().bool_type().into())
-    }
-
-    /// An integer type.
-    ///
-    /// :param int width: The number of bits in the integers.
-    /// :returns: The integer type.
-    /// :rtype: Type
-    #[pyo3(text_signature = "(width)")]
-    fn int(&self, py: Python, width: u32) -> PyResult<PyObject> {
-        self.create_type(py, |m| m.get_context().custom_width_int_type(width).into())
-    }
-
-    /// The double type.
-    ///
-    /// :type: Type
-    #[getter]
-    fn double(&self, py: Python) -> PyResult<PyObject> {
-        self.create_type(py, |m| m.get_context().f64_type().into())
-    }
-
-    /// The qubit type.
-    ///
-    /// :type: Type
-    #[getter]
-    fn qubit(&self, py: Python) -> PyResult<PyObject> {
-        self.create_type(py, |m| types::qubit(m).into())
-    }
-
-    /// The measurement result type.
-    ///
-    /// :type: Type
-    #[getter]
-    fn result(&self, py: Python) -> PyResult<PyObject> {
-        self.create_type(py, |m| types::result(m).into())
-    }
-
-    /// A function type.
-    ///
-    /// :param Type ret: The return type.
-    /// :param List[Type] params: The parameter types.
-    /// :returns: The function type.
-    /// :rtype: Type
-    #[staticmethod]
-    #[pyo3(text_signature = "(ret, params)")]
-    #[allow(clippy::needless_pass_by_value)]
-    fn function(py: Python, ret: &Type, params: Vec<Py<Type>>) -> PyResult<PyObject> {
-        context::require_same(
-            py,
-            params
-                .iter()
-                .map(|t| t.borrow(py).context().clone())
-                .chain([ret.context().clone()]),
-        )?;
-
-        let ty = function_type(
-            &ret.get(),
-            params.iter().map(|t| unsafe {
-                transmute::<AnyTypeEnum<'_>, AnyTypeEnum<'static>>(t.borrow(py).get())
-            }),
-        )
-        .ok_or_else(|| PyValueError::new_err("Invalid return or parameter type."))?;
-
-        unsafe { Type::from_any(py, ret.context().clone(), ty.into()) }
-    }
-}
-
-impl TypeFactory {
-    fn create_type(
-        &self,
-        py: Python,
-        f: impl for<'ctx> Fn(&InkwellModule<'ctx>) -> AnyTypeEnum<'ctx>,
-    ) -> PyResult<PyObject> {
-        let module = self.module.borrow(py);
-        let context = module.context.clone();
-        let ty = f(&module.module);
-        unsafe { Type::from_any(py, context, ty) }
+    pub(crate) fn context(&self) -> &Py<Context> {
+        &self.context
     }
 }
 
@@ -229,7 +133,7 @@ pub(crate) struct Builder {
 }
 
 impl Builder {
-    fn new(py: Python, module: Py<Module>) -> Self {
+    pub(crate) fn new(py: Python, module: Py<Module>) -> Self {
         let context = module.borrow(py).context.clone();
         let builder = {
             let context = context.borrow(py);
@@ -242,6 +146,10 @@ impl Builder {
             context,
             module,
         }
+    }
+
+    pub(crate) fn get(&self) -> &InkwellBuilder<'static> {
+        &self.builder
     }
 }
 
@@ -466,146 +374,6 @@ impl Builder {
             |_| call_if_some(r#true),
             |_| call_if_some(r#false),
         )
-    }
-}
-
-/// A simple module represents an executable program with these restrictions:
-///
-/// - There is one global qubit register and one global result register. Both are statically
-///   allocated with a fixed size.
-/// - There is only a single function that runs as the entry point.
-///
-/// :param str name: The name of the module.
-/// :param int num_qubits: The number of statically allocated qubits.
-/// :param int num_results: The number of statically allocated results.
-#[pyclass(unsendable)]
-#[pyo3(text_signature = "(name, num_qubits, num_results)")]
-pub(crate) struct SimpleModule {
-    module: Py<Module>,
-    builder: Py<Builder>,
-    types: Py<TypeFactory>,
-    num_qubits: u64,
-    num_results: u64,
-}
-
-#[pymethods]
-impl SimpleModule {
-    #[new]
-    fn new(py: Python, name: &str, num_qubits: u64, num_results: u64) -> PyResult<SimpleModule> {
-        let context = Py::new(py, Context::new(InkwellContext::create()))?;
-        let module = Py::new(py, Module::new(py, context, name))?;
-        let builder = Py::new(py, Builder::new(py, module.clone()))?;
-
-        {
-            let builder = builder.borrow(py);
-            let module = module.borrow(py);
-            module::simple_init(&module.module, &builder.builder, num_qubits, num_results);
-        }
-
-        let types = Py::new(
-            py,
-            TypeFactory {
-                module: module.clone(),
-            },
-        )?;
-
-        Ok(SimpleModule {
-            module,
-            builder,
-            types,
-            num_qubits,
-            num_results,
-        })
-    }
-
-    #[getter]
-    fn types(&self) -> Py<TypeFactory> {
-        self.types.clone()
-    }
-
-    /// The global qubit register.
-    ///
-    /// :type: Tuple[Value, ...]
-    #[getter]
-    fn qubits(&self, py: Python) -> PyResult<Vec<PyObject>> {
-        let builder = self.builder.borrow(py);
-        let module = self.module.borrow(py);
-        let builder = qirlib::Builder::from(&builder.builder, &module.module);
-        (0..self.num_qubits)
-            .map(|id| unsafe {
-                Value::from_any(py, module.context.clone(), builder.build_qubit(id))
-            })
-            .collect()
-    }
-
-    /// The global result register.
-    ///
-    /// :type: Tuple[Value, ...]
-    #[getter]
-    fn results(&self, py: Python) -> PyResult<Vec<PyObject>> {
-        let builder = self.builder.borrow(py);
-        let module = self.module.borrow(py);
-        let builder = qirlib::Builder::from(&builder.builder, &module.module);
-        (0..self.num_results)
-            .map(|id| unsafe {
-                Value::from_any(py, module.context.clone(), builder.build_result(id))
-            })
-            .collect()
-    }
-
-    /// The instruction builder.
-    ///
-    /// :type: Builder
-    #[getter]
-    fn builder(&self) -> Py<Builder> {
-        self.builder.clone()
-    }
-
-    /// Emits the LLVM IR for the module as plain text.
-    ///
-    /// :rtype: str
-    fn ir(&self, py: Python) -> PyResult<String> {
-        self.emit(py, |m| m.print_to_string().to_string())
-    }
-
-    /// Emits the LLVM bitcode for the module as a sequence of bytes.
-    ///
-    /// :rtype: bytes
-    fn bitcode<'py>(&self, py: Python<'py>) -> PyResult<&'py PyBytes> {
-        self.emit(py, |m| {
-            PyBytes::new(py, m.write_bitcode_to_memory().as_slice())
-        })
-    }
-
-    /// Adds a declaration for an externally linked function to the module.
-    ///
-    /// :param str name: The name of the function.
-    /// :param Type ty: The type of the function.
-    /// :return: The function value.
-    /// :rtype: Function
-    #[pyo3(text_signature = "(self, name, ty)")]
-    fn add_external_function(&mut self, py: Python, name: &str, ty: &Type) -> PyResult<PyObject> {
-        let module = self.module.borrow(py);
-        context::require_same(py, [&module.context, ty.context()])?;
-
-        let context = ty.context().clone();
-        let ty = unsafe { transmute::<AnyTypeEnum<'_>, AnyTypeEnum<'static>>(ty.get()) }
-            .into_function_type();
-        let function = module.module.add_function(name, ty, None);
-        unsafe { Value::from_any(py, context, function) }
-    }
-}
-
-impl SimpleModule {
-    fn emit<T>(&self, py: Python, f: impl Fn(&InkwellModule) -> T) -> PyResult<T> {
-        let module = self.module.borrow(py);
-        let builder = self.builder.borrow(py);
-        let ret = builder.builder.build_return(None);
-        let new_context = InkwellContext::create();
-        let new_module = clone_module(&module.module, &new_context)?;
-        ret.erase_from_basic_block();
-        module::simple_finalize(&new_module).map_err(PyOSError::new_err)?;
-        Ok(f(&new_module))
     }
 }
